@dscho 2020-04-18
> I'd like to submit this directly upstream via GGG after we've had a
> chance to review it and avoid the technical debt.

@jeffhostetler I think we should do both. Are you _totally_ opposed to having it in Git for Windows early? It's such a trivial patch, it will be simply skipped from the rebase once it arrives in any official Git version.