@jeffhostetler 2020-04-18
@dscho I’m ok with including in both. i was just trying to not cause you more work later.  i’ll update and merge this and send one to GGG and the list on monday.  thanks.