@derrickstolee 2020-04-20
@jeffhostetler any reason you didn't create a PR against `vfs-2.26.1` in microsoft/git? I could generate an installer and run it through some perf tests that way.