@jeffhostetler 2020-04-20
@derrickstolee  I'll merge this into GFW as is.  Then cherry-pick it into vfs so we can test it there.  And then send a patch to the mailing list via GGG.